Reorg of SprayProps
